refactor(startup): split orchestrator types
